requirements.txt
pegasus-wms.api
